backend/app/services/environmental_interaction.py
from typing import Dict, List, Any, Optional
from datetime import datetime
import logging
import json
import asyncio
import aiohttp
import requests
from sqlalchemy.orm import Session
from ..services.memory_service import Memory, get_db
from ..services.intelligence_service import IntelligenceEngine
from ..services.self_modification import SelfModificationEngine

logger = logging.getLogger(__name__)

class EnvironmentalInteractionEngine:
    def __init__(self):
        self.intelligence_engine = IntelligenceEngine()
        self.self_mod_engine = SelfModificationEngine()
        self.confidence_threshold = 0.85
        self.max_concurrent_requests = 5
        self.request_timeout = 30
        self.retry_attempts = 3
        
    async def research_topic(self, topic: str, depth: int = 3, db: Session = None) -> Dict[str, Any]:
        """Conduct autonomous research on a given topic."""
        try:
            # Initialize research session
            research_session = await self._initialize_research_session(topic)
            
            # Gather information from multiple sources
            sources = await self._gather_information(topic, depth)
            
            # Analyze and synthesize findings
            analysis = await self._analyze_findings(sources)
            
            # Generate comprehensive report
            report = await self._generate_research_report(analysis)
            
            # Store research results
            if db:
                self._store_research_results(report, db)
            
            return report
        except Exception as e:
            logger.error(f"Error conducting research: {str(e)}")
            raise

    async def _initialize_research_session(self, topic: str) -> Dict[str, Any]:
        """Initialize a new research session with appropriate parameters."""
        return {
            'topic': topic,
            'start_time': datetime.utcnow().isoformat(),
            'sources': [],
            'findings': [],
            'confidence': 0.0
        }

    async def _gather_information(self, topic: str, depth: int) -> List[Dict[str, Any]]:
        """Gather information from multiple sources."""
        sources = []
        
        # Research APIs to query
        research_apis = [
            'scholar',
            'news',
            'financial',
            'technical'
        ]
        
        async with aiohttp.ClientSession() as session:
            tasks = []
            for api in research_apis:
                task = self._query_research_api(session, api, topic)
                tasks.append(task)
            
            results = await asyncio.gather(*tasks, return_exceptions=True)
            
            for result in results:
                if isinstance(result, dict):
                    sources.append(result)
        
        return sources

    async def _query_research_api(self, session: aiohttp.ClientSession, api: str, topic: str) -> Dict[str, Any]:
        """Query a specific research API."""
        try:
            # This would be replaced with actual API calls
            return {
                'api': api,
                'topic': topic,
                'results': [],
                'confidence': 0.8
            }
        except Exception as e:
            logger.error(f"Error querying {api} API: {str(e)}")
            return {'api': api, 'error': str(e)}

    async def _analyze_findings(self, sources: List[Dict[str, Any]]) -> Dict[str, Any]:
        """Analyze and synthesize findings from multiple sources."""
        analysis = {
            'key_insights': [],
            'confidence': 0.0,
            'sources_analyzed': len(sources),
            'timestamp': datetime.utcnow().isoformat()
        }
        
        # Process findings from each source
        for source in sources:
            if 'results' in source:
                insights = await self._extract_insights(source['results'])
                analysis['key_insights'].extend(insights)
        
        # Calculate overall confidence
        analysis['confidence'] = self._calculate_analysis_confidence(analysis)
        
        return analysis

    async def _extract_insights(self, results: List[Any]) -> List[Dict[str, Any]]:
        """Extract key insights from research results."""
        insights = []
        
        # Process results to extract insights
        for result in results:
            insight = {
                'content': str(result),
                'confidence': 0.8,
                'timestamp': datetime.utcnow().isoformat()
            }
            insights.append(insight)
        
        return insights

    def _calculate_analysis_confidence(self, analysis: Dict[str, Any]) -> float:
        """Calculate confidence level for the analysis."""
        if not analysis['key_insights']:
            return 0.0
            
        confidence_scores = [insight.get('confidence', 0.0) for insight in analysis['key_insights']]
        return sum(confidence_scores) / len(confidence_scores)

    async def _generate_research_report(self, analysis: Dict[str, Any]) -> Dict[str, Any]:
        """Generate a comprehensive research report."""
        report = {
            'title': f"Research Report: {analysis.get('topic', 'Unknown Topic')}",
            'timestamp': datetime.utcnow().isoformat(),
            'analysis': analysis,
            'recommendations': await self._generate_recommendations(analysis),
            'confidence': analysis['confidence']
        }
        
        return report

    async def _generate_recommendations(self, analysis: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Generate recommendations based on research findings."""
        recommendations = []
        
        # Process insights to generate recommendations
        for insight in analysis['key_insights']:
            recommendation = {
                'action': f"Consider: {insight['content']}",
                'confidence': insight['confidence'],
                'priority': 'high' if insight['confidence'] > 0.8 else 'medium'
            }
            recommendations.append(recommendation)
        
        return recommendations

    async def interact_with_api(self, api_name: str, action: str, params: Dict[str, Any], db: Session = None) -> Dict[str, Any]:
        """Interact with external APIs."""
        try:
            # Validate API interaction
            if not self._validate_api_interaction(api_name, action, params):
                raise ValueError("Invalid API interaction parameters")
            
            # Execute API interaction
            result = await self._execute_api_interaction(api_name, action, params)
            
            # Store interaction result
            if db:
                self._store_api_interaction(result, db)
            
            return result
        except Exception as e:
            logger.error(f"Error interacting with API: {str(e)}")
            raise

    def _validate_api_interaction(self, api_name: str, action: str, params: Dict[str, Any]) -> bool:
        """Validate API interaction parameters."""
        # Add validation logic here
        return True

    async def _execute_api_interaction(self, api_name: str, action: str, params: Dict[str, Any]) -> Dict[str, Any]:
        """Execute API interaction."""
        # This would be replaced with actual API calls
        return {
            'api': api_name,
            'action': action,
            'result': 'success',
            'timestamp': datetime.utcnow().isoformat()
        }

    async def generate_document(self, document_type: str, content: Dict[str, Any], db: Session = None) -> Dict[str, Any]:
        """Generate a document based on provided content."""
        try:
            # Generate document
            document = await self._create_document(document_type, content)
            
            # Store document
            if db:
                self._store_document(document, db)
            
            return document
        except Exception as e:
            logger.error(f"Error generating document: {str(e)}")
            raise

    async def _create_document(self, document_type: str, content: Dict[str, Any]) -> Dict[str, Any]:
        """Create a document of the specified type."""
        document = {
            'type': document_type,
            'content': content,
            'timestamp': datetime.utcnow().isoformat(),
            'format': 'markdown'
        }
        
        # Add document-specific processing here
        
        return document

    def _store_research_results(self, report: Dict[str, Any], db: Session) -> None:
        """Store research results in memory."""
        memory = Memory(
            content=f"Research Report: {json.dumps(report)}",
            metadata={
                'type': 'research_report',
                'timestamp': datetime.utcnow().isoformat(),
                'details': report
            }
        )
        db.add(memory)
        db.commit()

    def _store_api_interaction(self, interaction: Dict[str, Any], db: Session) -> None:
        """Store API interaction result in memory."""
        memory = Memory(
            content=f"API Interaction: {json.dumps(interaction)}",
            metadata={
                'type': 'api_interaction',
                'timestamp': datetime.utcnow().isoformat(),
                'details': interaction
            }
        )
        db.add(memory)
        db.commit()

    def _store_document(self, document: Dict[str, Any], db: Session) -> None:
        """Store generated document in memory."""
        memory = Memory(
            content=f"Generated Document: {json.dumps(document)}",
            metadata={
                'type': 'generated_document',
                'timestamp': datetime.utcnow().isoformat(),
                'details': document
            }
        )
        db.add(memory)
        db.commit() 