backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.6.4
clickhouse-driver==0.2.6

# AI and Machine Learning
openai==1.3.0
anthropic==0.5.0
cohere==4.37
perplexity==0.1.0
replicate==0.22.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# WebSocket and Real-time
websockets==12.0
aiohttp==3.9.1
httpx==0.25.1

# Security
cryptography==41.0.5
pyjwt==2.8.0
bcrypt==4.0.1

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings==0.23.0

# Financial Dependencies
yfinance==0.2.33
pandas==2.1.3
numpy==1.26.2
ta==0.10.2

# Creative Expression
pillow==10.1.0
moviepy==1.0.3
pydub==0.25.1
soundfile==0.12.1

# Environmental Interaction
python-telegram-bot==20.7
tweepy==4.14.0
discord.py==2.3.2
slack-sdk==3.26.1

# GIS Dependencies
geopy==2.4.1
folium==0.14.0
shapely==2.0.2

# Social Network Dependencies
networkx==3.2.1
community==0.15
scipy==1.11.4

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinx-autodoc-typehints==1.25.2